add count to tab, cleanup
