--- embedded_gems/pathogen/app/components/pathogen/tab.rb
@@ -1,46 +1,0 @@
-# frozen_string_literal: true
-
-module Pathogen
-  class Tab < Pathogen::Component
-    TAG_DEFAULT = :a
-
-    WRAPPER_CLASSES = 'inline-flex items-center justify-center mr-2'
-    COUNT_CLASSES = 'ml-1 text-sm text-slate-600 dark:text-slate-400 rounded-full bg-slate-100 dark:bg-slate-900 inline-block p-0.5'
-    attr_reader :selected
-
-    def initialize(controls:, count: nil, selected: false, icon_classes: '', text: '',
-                   count_arguments: {}, wrapper_arguments: {}, **system_arguments)
-      @controls = controls
-      @count = count
-      @selected = selected
-      @icon_classes = icon_classes
-      @text = text
-
-      @system_arguments = system_arguments
-      @wrapper_arguments = wrapper_arguments
-      @count_arguments = count_arguments
-
-      @system_arguments[:tag] = TAG_DEFAULT
-
-      @wrapper_arguments[:tag] = :li
-      @wrapper_arguments[:classes] = WRAPPER_CLASSES
-
-      @system_arguments[:'aria-current'] = @selected ? 'page' : 'false'
-      @system_arguments[:classes] = generate_link_classes
-      @system_arguments[:'aria-controls'] = @controls
-
-      @count_arguments[:tag] = :span
-      @count_arguments[:classes] = COUNT_CLASSES
-    end
-
-    def generate_link_classes
-      if @selected
-        'inline-block p-4 border-b-2 rounded-t-lg text-primary-700
-        border-primary-700 active dark:text-primary-500 dark:border-primary-500'
-      else
-        'inline-block p-4 border-b-2 rounded-t-lg border-transparent
-        hover:text-slate-600 hover:border-slate-300 dark:hover:text-slate-300'
-      end
-    end
-  end
-end
