+++ embedded_gems/pathogen/app/components/pathogen/tabs_panel/count.html.erb
@@ -1,0 +1,3 @@
+<%= render Pathogen::BaseComponent.new(**@system_arguments) do %>
+<%= @count %>
+<% end %>
