+++ embedded_gems/pathogen/app/components/pathogen/tabs_panel/tab.html.erb
@@ -1,12 +1,9 @@
 <%= render Pathogen::BaseComponent.new(**@wrapper_arguments) do %>
 <%= render Pathogen::BaseComponent.new(**@system_arguments) do %>
+<%= icon %>
 <span>
   <%= @text %>
 </span>
-<% if @count %>
-<%= render Pathogen::BaseComponent.new(**@count_arguments) do %>
-<%= @count %>
-<% end %>
-<% end %>
+<%= count %>
 <% end %>
 <% end %>
