+++ test/components/previews/pathogen_tabs_panel_preview/with_count.html.erb
@@ -1,5 +1,11 @@
 <%= render Pathogen::TabsPanel.new(id: "default-tabs-panel-id", label: "default tabs panel") do |tabs| %>
-<%= tabs.with_tab(selected: true, href: "#", text: "Tab 1", count: 1) %>
-<%= tabs.with_tab(selected: false, href: "#", text: "Tab 2", count: 10) %>
-<%= tabs.with_tab(selected: false, href: "#", text: "Tab 3", count: 100) %>
+<%= tabs.with_tab(selected: true, href: "#", text: "Tab 1") do |tab| %>
+<%= tab.with_count(count: 1) %>
+<% end %>
+<%= tabs.with_tab(selected: false, href: "#", text: "Tab 2") do |tab| %>
+<%= tab.with_count(count: 10) %>
+<% end %>
+<%= tabs.with_tab(selected: false, href: "#", text: "Tab 3") do |tab| %>
+<%= tab.with_count(count: 100) %>
+<% end %>
 <% end %>
